Add controller for redirecting to Nosto url
