+++ view/adminhtml/templates/connection.phtml
@@ -42,30 +42,28 @@ use Nosto\Tagging\Block\Adminhtml\Account\Connection;
  */
 ?>
 
-<?php if ($block->getAccount()): ?>
-    <?php if ($block->getAccount()->isConnectedToNosto()): ?>
-        <button id="nosto_open"
-                title="Open"
-                type="button"
-                class="action-default primary"
-                onclick="window.open('<?= $block->escapeUrl($block->getNostoUrl()) ?>', '_blank');">
-            <span>Open Nosto Dashboard</span>
-        </button>
+<?php if ($block->nostoInstalledAndEnabled()): ?>
+    <button id="nosto_open"
+            title="Open"
+            type="button"
+            class="action-default primary"
+            onclick="window.open('<?= $block->escapeUrl($block->redirectToNostoUrl()) ?>', '_blank');">
+        <span>Open Nosto Dashboard</span>
+    </button>
 
-        <button id="nosto_remove"
-                title="Remove"
-                type="button"
-                class="action-default"
-                onclick="window.open('<?= $block->escapeUrl($block->getAccountDeleteUrl()) ?>', '_self');">
-            <span>Remove Account</span>
-        </button>
-    <?php endif; ?>
+    <button id="nosto_remove"
+            title="Remove"
+            type="button"
+            class="action-default"
+            onclick="window.open('<?= $block->escapeUrl($block->getAccountDeleteUrl()) ?>', '_self');">
+        <span>Remove Account</span>
+    </button>
 <?php else: ?>
     <button id="nosto_connect"
             title="Connect"
             type="button"
             class="action-default primary"
-            onclick="window.open('<?= $block->escapeUrl($block->getNostoUrl()) ?>', '_self');">
+            onclick="window.open('<?= $block->escapeUrl($block->redirectToNostoUrl()) ?>', '_self');">
         <span>Connect to Nosto</span>
     </button>
 <?php endif; ?>
